requirements.txt
Django>=5.0
psycopg2-binary>=2.9 
requests
beautifulsoup4
djangorestframework
django-filter

